requirements-test.txt
immunity-utils[qa,selenium] @ https://github.com/edge-servers/immunity-utils/tarball/master
django-cors-headers~=4.0.0
django-redis~=5.2.0
channels_redis~=4.1.0
pytest-asyncio~=0.21.0
pytest-django~=4.5.0
freezegun~=1.2.2
